Implement change team name on team settings.
